Fix image version command group name
