+++ src/devcenter/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+0.1.1
+* Fix wrong command group name
+* Add correct author email
+
 0.1.0
 ++++++
 * Initial release.
