+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/image_version/__cmd_group.py
@@ -12,7 +12,7 @@ from azure.cli.core.aaz import *
 
 
 @register_command_group(
-    "devcenter admin image-verion",
+    "devcenter admin image-version",
     is_preview=True,
 )
 class __CMDGroup(AAZCommandGroup):
